LLM & Generative AI/pyproject.toml
[project]
name = "genai-pipeline"
version = "0.1.0"
description = "LLM & Generative AI pipeline với RAG, Agents, và Tools"
requires-python = ">=3.10"
dependencies = [
    # LLM Clients
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "groq>=0.6.0",

    # Vector DB & Embeddings
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",

    # Data processing
    "pandas>=2.2.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "tiktoken>=0.7.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "pre-commit>=4.0.0",
]
local = [
    "ollama>=0.3.0",
    "llama-index>=0.11.0",
]
